requirements.txt
# # DO NOT include azure-functions-worker in this file
# # The Python Worker is managed by Azure Functions platform
# # Manually managing azure-functions-worker may cause unexpected issues

# azure-functions
# joblib==1.0.1

# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob==2.1.0
absl-py==1.0.0
astor==0.8.1
gast==0.3.3
grpcio>=1.32.0
grpcio-tools==1.32.0
grpcio-health-checking==1.32.0
grpcio-reflection==1.32.0
importlib-metadata==1.7.0
joblib==1.1.0
Keras==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
mock==4.0.2
numpy==1.23.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
pathlib==1.0.1
pandas==1.1.5
protobuf==3.13.0
scipy==1.9.3
six==1.15.0
tensorboard==2.9.0
tensorflow==2.9.0
tensorflow-estimator==2.9.0
termcolor==1.1.0
Werkzeug==1.0.1
zipp==3.1.0
applicationinsights==0.11.9
scikit-image==0.19.3
scikit-learn >= 0.24.1
h5py==3.7.0
xgboost==1.5.2